+++ frontend/frontend/components/auth/LoggedInUser/LoggedInUser.css
@@ -1,0 +1,10 @@
+.logged-in-info {
+    position: fixed;
+    top: 0.625rem; /* 10px */
+    right: 0.625rem; /* 10px */
+    background-color: rgba(33, 33, 33, 0.8);
+    padding: 0.3125rem 0.625rem; /* 5px 10px */
+    border-radius: 0.3125rem; /* 5px */
+    z-index: 9999;
+    font-size: 0.75rem; /* 12px */
+  }
