+++ frontend/frontend/components/buttons/copy/CopyToClipboardButton.css
