+++ frontend/frontend/components/index.tsx
@@ -1,10 +1,10 @@
 // Auth Components
-export { default as ICWalletList } from './auth/ICWalletList';
-export { default as LoggedInUser } from './auth/LoggedInUser';
+export { default as ICWalletList } from './auth/ICWalletList/ICWalletList';
+export { default as LoggedInUser } from './auth/LoggedInUser/LoggedInUser';
 
 // Upload Components
-export { default as UploadButton } from './upload/UploadButton';
-export { default as DragAndDropContainer } from './upload/DragAndDropContainer';
+export { default as UploadButton } from './buttons/upload/UploadAssetButton';
+export { default as DragAndDropContainer } from './overlays/DragAndDropOverlay/DragAndDropOverlay';
 
 // Dropdown Components
 export { default as SettingsDropdown } from './dropdowns/SettingsDropdown';
@@ -13,8 +13,8 @@ export { default as SettingsDropdown } from './dropdowns/SettingsDropdown';
 export { default as ViewToggle } from './menus/ViewToggle';
 
 // Cycles Components
-export { default as CyclesTopUpComponent } from './cycles/CyclesTopUpButton';
-export { default as GetBalancesComponent } from './cycles/GetBalancesButton';
+export { default as CyclesTopUpComponent } from './buttons/cycles/CyclesTopUpButton';
+export { default as GetBalancesComponent } from './buttons/cycles/GetBalancesButton';
 
 // Overlay Components
 export { default as LoadingOverlay } from './overlays/LoadingOverlay/LoadingOverlay';
@@ -22,4 +22,4 @@ export { default as ErrorNotification } from './overlays/ErrorNotification/Error
 export { default as DeleteConfirmation } from './overlays/DeleteConfirmation/DeleteConfirmation';
 
 // Copy Components
-export { default as CopyToClipboard } from './copy/CopyToClipboard';
+export { default as CopyToClipboard } from './buttons/copy/CopyToClipboardButton';
